Bump dependency versions installed by pip.
This now matches the `requirements.txt` file in the current master repo, apart from `mapbox-vector-tile`, which has been kept as an overridable git branch/tag/commit.

This is needed because the master branch of `tilequeue` has been updated to the latest version of the EDTF library, which has a slightly different interface to the raw git commit that we were using here before.